task 2

+++ 0x00-python_variable_annotations/2-floor.py
@@ -2,12 +2,12 @@
 """ function floor which takes a float n as argument
     and returns the floor of the float.
 """
+import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """
     :param n: a float number
     :return: the floor of the float
     """
-    import math
     return math.floor(n)
